refactor(frontend): split ModsTab/FilesTab/BackupsTab and socketClient into domain modules
