+++ panel/frontend/src/lib/services/socketClient.ts
@@ -1,107 +1,27 @@
 import { io, type Socket } from 'socket.io-client';
 import { get, writable } from 'svelte/store';
-import { _ } from 'svelte-i18n';
 import { isAuthenticated } from '$lib/stores/auth';
 import { panelConfig } from '$lib/stores/config';
-import {
-  addLog,
-  clearLogs,
-  extractTimestamp,
-  formatTimestamp,
-  getLogType,
-  hasMoreHistory,
-  initialLoadDone,
-  isLoadingMore,
-  loadedCount,
-  logs,
-  prependLogs
-} from '$lib/stores/console';
-import { currentPath, fileList, isFilesLoading, setEditorContent, setEditorStatus } from '$lib/stores/files';
-import {
-  apiConfigured,
-  availableUpdates,
-  cfApiConfigured,
-  currentPage,
-  curseforgeStatus,
-  hasMore,
-  installedMods,
-  isModsLoading,
-  modtaleStatus,
-  searchResults,
-  total
-} from '$lib/stores/mods';
+import { clearLogs, hasMoreHistory, initialLoadDone, loadedCount } from '$lib/stores/console';
+import { currentPath, fileList } from '$lib/stores/files';
+import { installedMods } from '$lib/stores/mods';
 import { currentRoute, navigateToDashboard, navigateToServer } from '$lib/stores/router';
-import {
-  downloaderAuth,
-  downloadProgress,
-  filesReady,
-  isCheckingFiles,
-  serverStatus,
-  updateInfo
-} from '$lib/stores/server';
-import type { Server } from '$lib/stores/servers';
-import { activeServerId, servers, updateServerStatus } from '$lib/stores/servers';
+import { downloaderAuth, filesReady, isCheckingFiles, serverStatus } from '$lib/stores/server';
 import { showToast } from '$lib/stores/ui';
-import type {
-  ActionStatus,
-  CommandResult,
-  DownloadStatusEvent,
-  FileListResult,
-  FileOperationResult,
-  FileReadResult,
-  FileSaveResult,
-  FilesReady,
-  InstalledMod,
-  LogEntry,
-  LogsHistoryData,
-  ModOperationResult,
-  ModProject,
-  ModSearchResult,
-  ModUpdatesResult,
-  ServerStatus,
-  UncheckedMod
-} from '$lib/types';
+import { registerConsoleEvents } from './socket/consoleEvents';
+import { registerDownloadEvents, resetDownloadProgress } from './socket/downloadEvents';
+import { registerFilesEvents } from './socket/filesEvents';
+import { registerModsEvents } from './socket/modsEvents';
+import { registerServerEvents } from './socket/serverEvents';
+import { registerUpdateEvents } from './socket/updateEvents';
 
 export const socket = writable<Socket | null>(null);
 export const isConnected = writable<boolean>(false);
 export const joinedServerId = writable<string | null>(null);
 
 let socketInstance: Socket | null = null;
-let dlStartTime: number | null = null;
-let dlTimer: ReturnType<typeof setInterval> | null = null;
 let routeUnsubscribe: (() => void) | null = null;
 let lastJoinedServerId: string | null = null;
-
-type Translate = (key: string, options?: { values?: Record<string, string | number> }) => string;
-
-function getUncheckedModReasonLabel(t: Translate, reason: UncheckedMod['reason']): string {
-  switch (reason) {
-    case 'local_mod':
-      return t('uncheckedReasonLocalMod');
-    case 'missing_project_id':
-      return t('uncheckedReasonMissingProjectId');
-    case 'modtale_api_not_configured':
-      return t('uncheckedReasonModtaleApi');
-    case 'curseforge_api_not_configured':
-      return t('uncheckedReasonCurseforgeApi');
-    case 'unknown_provider':
-      return t('uncheckedReasonUnknownProvider');
-    case 'provider_lookup_failed':
-      return t('uncheckedReasonLookupFailed');
-  }
-}
-
-function formatUncheckedModsMessage(t: Translate, uncheckedMods: UncheckedMod[]): string {
-  const details = uncheckedMods
-    .map((mod) => `${mod.projectTitle}: ${getUncheckedModReasonLabel(t, mod.reason)}`)
-    .join('; ');
-  return t('noUpdatesUncheckedDetails', {
-    values: {
-      count: uncheckedMods.length,
-      details
-    }
-  });
-}
 
 export function connectSocket(): Socket {
   if (socketInstance?.connected) {
@@ -163,408 +83,12 @@ export function connectSocket(): Socket {
     }
   });
 
-  // Server join result
-  socketInstance.on('server:joined', ({ serverId, server }: { serverId: string; server?: Server }) => {
-    joinedServerId.set(serverId);
-    // Update servers store with fresh data from backend
-    if (server) {
-      servers.update((list) => {
-        const idx = list.findIndex((s) => s.id === serverId);
-        if (idx >= 0) {
-          list[idx] = { ...list[idx], ...server };
-          return [...list];
-        }
-        return [...list, server];
-      });
-    }
-    // Request file list - always available via local filesystem
-    socketInstance?.emit('files:list', '/');
-  });
-
-  socketInstance.on('server:join-error', ({ error }: { error: string }) => {
-    showToast(`Error: ${error}`, 'error');
-  });
-
-  // Server status
-  socketInstance.on('status', (s: ServerStatus) => {
-    const wasRunning = get(serverStatus).running;
-    const isNowRunning = s.running;
-
-    serverStatus.set({
-      running: isNowRunning,
-      status: s.status || 'unknown',
-      startedAt: s.startedAt
-    });
-
-    const serverId = get(activeServerId);
-    if (serverId) {
-      updateServerStatus(serverId, isNowRunning ? 'running' : 'stopped');
-    }
-
-    if (isNowRunning && !wasRunning) {
-      socketInstance?.emit('files:list', '/');
-      socketInstance?.emit('mods:list');
-    }
-  });
-
-  socketInstance.on('files', (f: FilesReady) => {
-    filesReady.set({
-      hasJar: f.hasJar,
-      hasAssets: f.hasAssets,
-      ready: f.ready
-    });
-    isCheckingFiles.set(false);
-    socketInstance?.emit('mods:check-config');
-    socketInstance?.emit('cf:check-config');
-  });
-
-  // Downloader auth status
-  socketInstance.on('downloader-auth', (a: boolean) => {
-    downloaderAuth.set(a);
-    isCheckingFiles.set(false);
-  });
-
-  // Download status
-  socketInstance.on('download-status', handleDownloadStatus);
-
-  // Logs
-  socketInstance.on('log', (msg: string) => {
-    addLog(msg.trim());
-  });
-
-  socketInstance.on('logs:history', handleLogsHistory);
-
-  // Command result
-  socketInstance.on('command-result', (r: CommandResult) => {
-    if (r.error) {
-      addLog(`Error: ${r.error}`, 'error');
-      showToast(get(_)('failed'), 'error');
-    }
-  });
-
-  // Action status
-  socketInstance.on('action-status', (s: ActionStatus) => {
-    const t = get(_);
-    if (s.success) {
-      const msgs: Record<string, string> = {
-        restart: t('restarted'),
-        stop: t('stopped'),
-        start: t('started'),
-        wipe: t('dataWiped'),
-        kill: t('killed')
-      };
-      showToast(msgs[s.action] || t('done'));
-    } else if (s.error) {
-      showToast(`${s.action} ${t('failed')}`, 'error');
-    }
-  });
-
-  // Error
-  socketInstance.on('error', (m: string) => {
-    addLog(m, 'error');
-    showToast(m, 'error');
-  });
-
-  // File manager events
-  socketInstance.on('files:list-result', (result: FileListResult) => {
-    isFilesLoading.set(false);
-    if (result.success) {
-      currentPath.set(result.path);
-      fileList.set(result.files);
-    } else {
-      showToast(`Error: ${result.error}`, 'error');
-      fileList.set([]);
-    }
-  });
-
-  socketInstance.on('files:read-result', (result: FileReadResult) => {
-    if (result.success && result.content !== undefined) {
-      setEditorContent(result.content);
-    } else if (result.binary) {
-      setEditorStatus('Cannot edit binary file', 'error');
-    } else {
-      setEditorStatus(`Error: ${result.error}`, 'error');
-    }
-  });
-
-  socketInstance.on('files:save-result', (result: FileSaveResult) => {
-    if (result.success) {
-      setEditorStatus(result.backup?.success ? 'Saved (backup created)' : 'Saved', 'saved');
-      showToast('File saved');
-    } else {
-      setEditorStatus(`Error: ${result.error}`, 'error');
-      showToast('Save failed', 'error');
-    }
-  });
-
-  socketInstance.on('files:mkdir-result', (result: FileOperationResult) => {
-    if (result.success) {
-      showToast('Folder created');
-      emit('files:list', get(currentPath));
-    } else {
-      showToast(`Error: ${result.error}`, 'error');
-    }
-  });
-
-  socketInstance.on('files:delete-result', (result: FileOperationResult) => {
-    if (result.success) {
-      showToast('Deleted');
-      emit('files:list', get(currentPath));
-    } else {
-      showToast(`Error: ${result.error}`, 'error');
-    }
-  });
-
-  socketInstance.on('files:rename-result', (result: FileOperationResult) => {
-    if (result.success) {
-      showToast('Renamed');
-      emit('files:list', get(currentPath));
-    } else {
-      showToast(`Error: ${result.error}`, 'error');
-    }
-  });
-
-  // Mods events
-  socketInstance.on('mods:config-status', (result: { configured: boolean; valid: boolean; error?: string }) => {
-    modtaleStatus.set(result);
-    apiConfigured.set(result.configured && result.valid);
-    if (result.configured && !result.valid && result.error) {
-      showToast(`Modtale: ${result.error}`, 'error');
-    }
-  });
-
-  socketInstance.on('mods:list-result', (result: { success: boolean; mods?: InstalledMod[]; error?: string }) => {
-    isModsLoading.set(false);
-    if (result.success && result.mods) {
-      installedMods.set(result.mods);
-    } else {
-      showToast(`${get(_)('error')}: ${result.error}`, 'error');
-    }
-  });
-
-  socketInstance.on('mods:search-result', (result: ModSearchResult) => {
-    isModsLoading.set(false);
-    if (result.success) {
-      searchResults.set(result.projects || []);
-      total.set(result.total || 0);
-      hasMore.set(result.hasMore || false);
-      currentPage.set(result.page || 1);
-    } else {
-      showToast(`Error: ${result.error}`, 'error');
-      searchResults.set([]);
-    }
-  });
-
-  socketInstance.on('mods:install-result', (result: ModOperationResult) => {
-    const t = get(_);
-    if (result.success) {
-      showToast(t('modInstalled'));
-      emit('mods:list');
-    } else {
-      showToast(`${t('installFailed')}: ${result.error}`, 'error');
-    }
-  });
-
-  socketInstance.on('mods:uninstall-result', (result: ModOperationResult) => {
-    const t = get(_);
-    if (result.success) {
-      showToast(t('modUninstalled'));
-      emit('mods:list');
-    } else {
-      showToast(`${t('error')}: ${result.error}`, 'error');
-    }
-  });
-
-  socketInstance.on('mods:enable-result', (result: ModOperationResult) => {
-    const t = get(_);
-    if (result.success) {
-      showToast(t('modEnabled'));
-      emit('mods:list');
-    } else {
-      showToast(`${t('error')}: ${result.error}`, 'error');
-    }
-  });
-
-  socketInstance.on('mods:disable-result', (result: ModOperationResult) => {
-    const t = get(_);
-    if (result.success) {
-      showToast(t('modDisabled'));
-      emit('mods:list');
-    } else {
-      showToast(`${t('error')}: ${result.error}`, 'error');
-    }
-  });
-
-  socketInstance.on('mods:check-updates-result', (result: ModUpdatesResult) => {
-    const t = get(_);
-    if (result.success) {
-      availableUpdates.set(result.updates || []);
-      if (result.updates && result.updates.length > 0) {
-        showToast(t('updatesAvailable', { values: { count: result.updates.length } }));
-      } else if ((result.uncheckedMods || 0) > 0) {
-        showToast(formatUncheckedModsMessage(t, result.uncheckedModDetails || []), 'warning');
-      } else {
-        showToast(t('noUpdates'));
-      }
-    } else {
-      showToast(`${t('error')}: ${result.error}`, 'error');
-    }
-  });
-
-  socketInstance.on('mods:update-result', (result: ModOperationResult) => {
-    const t = get(_);
-    if (result.success) {
-      showToast(t('modUpdated'));
-      emit('mods:list');
-      emit('mods:check-updates');
-    } else {
-      showToast(`${t('updateFailed')}: ${result.error}`, 'error');
-    }
-  });
-
-  socketInstance.on('mods:get-result', (result: { success: boolean; project?: ModProject; error?: string }) => {
-    if (result.success && result.project) {
-      const project = result.project;
-      // Update search results with full project details
-      searchResults.update((results) => {
-        const index = results.findIndex((m) => m.id === project.id);
-        if (index >= 0) {
-          results[index] = project;
-        }
-        return results;
-      });
-      // Auto-install with version info now available
-      const latestVersion = project.latestVersion || project.versions?.[0];
-      if (latestVersion?.id) {
-        emit('mods:install', {
-          projectId: project.id,
-          versionId: latestVersion.id,
-          metadata: {
-            projectTitle: project.title,
-            projectSlug: project.slug,
-            projectIconUrl: project.iconUrl,
-            versionName: latestVersion.version,
-            classification: project.classification,
-            fileName: latestVersion.fileName
-          }
-        });
-      }
-    } else if (result.error) {
-      showToast(`${get(_)('error')}: ${result.error}`, 'error');
-    }
-  });
-
-  socketInstance.on('cf:config-status', (result: { configured: boolean; valid: boolean; error?: string }) => {
-    curseforgeStatus.set(result);
-    cfApiConfigured.set(result.configured && result.valid);
-    if (result.configured && !result.valid && result.error) {
-      showToast(`CurseForge: ${result.error}`, 'error');
-    }
-  });
-
-  socketInstance.on('cf:search-result', (result: ModSearchResult) => {
-    isModsLoading.set(false);
-    if (result.success) {
-      searchResults.set(result.projects || []);
-      total.set(result.total || 0);
-      hasMore.set(result.hasMore || false);
-      currentPage.set(result.page || 1);
-    } else {
-      showToast(`Error: ${result.error}`, 'error');
-      searchResults.set([]);
-    }
-  });
-
-  socketInstance.on('cf:get-result', (result: { success: boolean; project?: ModProject; error?: string }) => {
-    if (result.success && result.project) {
-      const project = result.project;
-      searchResults.update((results) => {
-        const index = results.findIndex((m) => m.id === project.id);
-        if (index >= 0) {
-          results[index] = project;
-        }
-        return results;
-      });
-      // Auto-install with version info now available
-      const latestVersion = project.latestVersion || project.versions?.[0];
-      if (latestVersion?.id) {
-        emit('cf:install', {
-          modId: project.id,
-          fileId: latestVersion.id,
-          metadata: {
-            projectTitle: project.title,
-            projectSlug: project.slug,
-            projectIconUrl: project.iconUrl,
-            versionName: latestVersion.version,
-            classification: project.classification,
-            fileName: latestVersion.fileName
-          }
-        });
-      }
-    } else if (result.error) {
-      showToast(`${get(_)('error')}: ${result.error}`, 'error');
-    }
-  });
-
-  socketInstance.on('cf:install-result', (result: ModOperationResult) => {
-    const t = get(_);
-    if (result.success) {
-      showToast(t('modInstalled'));
-      emit('mods:list');
-    } else {
-      showToast(`${t('installFailed')}: ${result.error}`, 'error');
-    }
-  });
-
-  // Server update events
-  socketInstance.on(
-    'update:check-result',
-    (result: {
-      success: boolean;
-      lastUpdate: string | null;
-      daysSinceUpdate: number | null;
-      code?: string;
-      error?: string;
-    }) => {
-      updateInfo.update((u) =>
-        result.success
-          ? {
-              ...u,
-              isChecking: false,
-              lastUpdate: result.lastUpdate,
-              daysSinceUpdate: result.daysSinceUpdate
-            }
-          : {
-              ...u,
-              isChecking: false
-            }
-      );
-
-      if (!result.success) {
-        const message =
-          result.code === 'CONTAINER_NOT_RUNNING' ? get(_)('downloadRequiresRunningServer') : result.error;
-        if (message) {
-          showToast(message, result.code === 'CONTAINER_NOT_RUNNING' ? 'warning' : 'error');
-        }
-      }
-    }
-  );
-
-  socketInstance.on('update:status', (data: { status: string; message: string }) => {
-    updateInfo.update((u) => ({
-      ...u,
-      updateStatus: data.message,
-      isUpdating: data.status !== 'complete' && data.status !== 'error'
-    }));
-
-    if (data.status === 'complete') {
-      showToast(get(_)('updateComplete'));
-      emit('check-files');
-      emit('update:check');
-    } else if (data.status === 'error') {
-      showToast(`${get(_)('updateFailed')}: ${data.message}`, 'error');
-    }
-  });
+  registerServerEvents(socketInstance);
+  registerConsoleEvents(socketInstance);
+  registerDownloadEvents(socketInstance);
+  registerFilesEvents(socketInstance);
+  registerModsEvents(socketInstance);
+  registerUpdateEvents(socketInstance);
 
   socket.set(socketInstance);
   return socketInstance;
@@ -606,17 +130,7 @@ export function joinServer(serverId: string): void {
     downloaderAuth.set(false);
     isCheckingFiles.set(true); // Mark as checking until we get response
 
-    // Reset download progress completely
-    stopDlTimer();
-    downloadProgress.set({
-      active: false,
-      status: '',
-      percentage: 0,
-      step: 'auth',
-      authUrl: null,
-      authCode: null,
-      time: '0s'
-    });
+    resetDownloadProgress();
 
     socketInstance.emit('server:join', serverId);
     lastJoinedServerId = serverId;
@@ -632,183 +146,5 @@ export function leaveServer(): void {
     lastJoinedServerId = null;
     // Navigate to dashboard (this also sets activeServerId to null via router)
     navigateToDashboard();
-  }
-}
-
-function handleDownloadStatus(d: DownloadStatusEvent & { serverId?: string }): void {
-  // Ignore events from other servers
-  const currentServerId = get(activeServerId);
-  if (d.serverId && d.serverId !== currentServerId) {
-    console.log('[Download] Ignoring event for different server:', d.serverId, 'current:', currentServerId);
-    return;
-  }
-
-  console.log('[Download] Status:', d.status, 'Message:', d.message);
-  switch (d.status) {
-    case 'starting':
-      dlStartTime = Date.now();
-      startDlTimer();
-      downloadProgress.set({
-        active: true,
-        status: get(_)('starting'),
-        percentage: 5,
-        step: 'auth',
-        authUrl: null,
-        authCode: null,
-        time: '0s'
-      });
-      break;
-
-    case 'auth-required':
-      if (d.message) {
-        addLog(d.message, 'auth');
-        const urlRegex = /https:\/\/oauth\.accounts\.hytale\.com\S+/;
-        const codeRegex = /(?:user_code[=:]\s*|code:\s*)([\w]+)/i;
-        const url = urlRegex.exec(d.message);
-        const code = codeRegex.exec(d.message);
-        downloadProgress.update((p) => ({
-          ...p,
-          status: get(_)('waitingAuth'),
-          percentage: 10,
-          authUrl: url ? url[0] : null,
-          authCode: code ? code[1] : null
-        }));
-      }
-      break;
-
-    case 'output':
-      if (d.message) {
-        addLog(d.message);
-        if (d.message.toLowerCase().includes('download')) {
-          downloadProgress.update((p) => ({
-            ...p,
-            status: get(_)('downloading'),
-            percentage: 30,
-            step: 'download'
-          }));
-        }
-        const pMatch = d.message.match(/(\d+)\.?\d*%/);
-        if (pMatch) {
-          const percent = Number.parseFloat(pMatch[0]);
-          downloadProgress.update((p) => ({
-            ...p,
-            status: `${Math.round(percent)}%`,
-            percentage: 30 + percent * 0.5
-          }));
-        }
-      }
-      break;
-
-    case 'extracting':
-      if (d.message) addLog(d.message, 'info');
-      downloadProgress.update((p) => ({
-        ...p,
-        status: get(_)('extracting'),
-        percentage: 85,
-        step: 'extract'
-      }));
-      break;
-
-    case 'complete':
-      stopDlTimer();
-      downloadProgress.update((p) => ({
-        ...p,
-        active: false,
-        status: get(_)('done'),
-        percentage: 100,
-        step: 'complete',
-        authUrl: null,
-        authCode: null
-      }));
-      showToast(get(_)('downloadComplete'));
-      addLog(get(_)('filesReadyMsg'), 'info');
-      break;
-
-    case 'done':
-    case 'error':
-      stopDlTimer();
-      if (d.status === 'error') {
-        downloadProgress.update((p) => ({
-          ...p,
-          status: get(_)('error'),
-          active: false
-        }));
-        if (d.code === 'CONTAINER_NOT_RUNNING') {
-          const message = get(_)('downloadRequiresRunningServer');
-          showToast(message, 'warning');
-          addLog(message, 'warn');
-        } else if (d.message) {
-          showToast(d.message, 'error');
-          addLog(`${get(_)('error')}: ${d.message}`, 'error');
-        }
-      } else {
-        downloadProgress.update((p) => ({ ...p, active: false }));
-      }
-      break;
-  }
-}
-
-function handleLogsHistory(data: LogsHistoryData): void {
-  if (data.error) {
-    console.error('Failed to load logs:', data.error);
-    isLoadingMore.set(false);
-    return;
-  }
-
-  if (!data.logs || data.logs.length === 0) {
-    hasMoreHistory.set(data.hasMore !== false);
-    isLoadingMore.set(false);
-    return;
-  }
-
-  if (data.initial) {
-    clearLogs();
-  }
-
-  const parsedLogs: LogEntry[] = data.logs
-    .map((line) => {
-      const ts = extractTimestamp(line);
-      const cleaned = line.replace(/^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}(?:\.\d+)?Z?\s*/, '').trim();
-      return {
-        text: cleaned,
-        type: getLogType(cleaned),
-        timestamp: formatTimestamp(ts)
-      };
-    })
-    .filter((l) => l.text);
-
-  loadedCount.update((c) => c + data.logs.length);
-
-  if (data.hasMore === false) {
-    hasMoreHistory.set(false);
-  }
-
-  if (data.initial) {
-    logs.set(parsedLogs);
-    setTimeout(() => initialLoadDone.set(true), 500);
-  } else {
-    prependLogs(parsedLogs);
-  }
-
-  isLoadingMore.set(false);
-}
-
-function formatSec(s: number): string {
-  return s < 60 ? `${s}s` : `${Math.floor(s / 60)}m ${s % 60}s`;
-}
-
-function startDlTimer(): void {
-  dlTimer = setInterval(() => {
-    if (dlStartTime) {
-      const elapsed = Math.floor((Date.now() - dlStartTime) / 1000);
-      downloadProgress.update((p) => ({ ...p, time: formatSec(elapsed) }));
-    }
-  }, 1000);
-}
-
-function stopDlTimer(): void {
-  if (dlTimer) {
-    clearInterval(dlTimer);
-    dlTimer = null;
   }
 }
